proyecto/backEnd/requirements.txt
requests
python-dotenv
pandas
scikit-learn
joblib
fastapi
uvicorn
pyarrow
cors